Update uint64 type mapping in convert.go

+++ pkg/converter/convert.go
@@ -21,7 +21,7 @@ var (
 		".google.protobuf.Int32Value":  "INTEGER",
 		".google.protobuf.Int64Value":  "INTEGER",
 		".google.protobuf.UInt32Value": "INTEGER",
-		".google.protobuf.UInt64Value": "INTEGER",
+		".google.protobuf.UInt64Value": "NUMERIC",
 		".google.protobuf.DoubleValue": "FLOAT",
 		".google.protobuf.FloatValue":  "FLOAT",
 		".google.protobuf.BoolValue":   "BOOLEAN",
@@ -35,7 +35,7 @@ var (
 		descriptor.FieldDescriptorProto_TYPE_FLOAT:  "FLOAT",
 
 		descriptor.FieldDescriptorProto_TYPE_INT64:    "INTEGER",
-		descriptor.FieldDescriptorProto_TYPE_UINT64:   "INTEGER",
+		descriptor.FieldDescriptorProto_TYPE_UINT64:   "NUMERIC",
 		descriptor.FieldDescriptorProto_TYPE_INT32:    "INTEGER",
 		descriptor.FieldDescriptorProto_TYPE_UINT32:   "INTEGER",
 		descriptor.FieldDescriptorProto_TYPE_FIXED64:  "INTEGER",
